Updated rng seed to use built-in random device when no seed is provided

+++ hdr/engine.hpp
@@ -79,7 +79,7 @@ class Metropolis
 	int minAdaptationLoops;
 	int maxAdaptationLoops;
 	bool rngSetSeed;
-	int rngSeed;
+	unsigned long int rngSeed;
 	EngineOutputLevel outputLevel;
 	STMOutput::OutputOptions posteriorOptions;
